+++ instagramz/src/App.js
@@ -7,6 +7,7 @@ import dummyData from './dummy-data';
 import PostContainer from './components/PostContainer/PostContainer';
 
 import FilterTest from './FilterTest';
+import PostPage from './components/PostContainer/PostsPage';
 
 
 
@@ -24,54 +25,12 @@ class App extends Component {
 
   }
 
-  componentDidMount() {
-    
-    this.setState (
-      {
-        content: dummyData,
-      }
-    )
-
-    console.log(this.state.content)
-  }
-
-
-  searching = (event, thing) => {
-    console.log(thing)
-    event.preventDefault();
-    this.setState (
-      {
-        search: thing
-      }
-    )
-   
-  }
-
-
-  filter = info => {
-    const data = this.state.content;
-    //const result = data.filter(search => search.username === this.state.search);
-    const result = data.filter(search => search.username.includes(this.state.search));
-
-    if (this.state.search == 'everything'){
-      return data;
-    }
-    else {
-      return result;
-    }
-
-  } 
 
   render() {
     return (
-      <div className="App">
-      <SearchBar searching={this.searching}/>
-      <PostContainer content={this.filter()} />
-            
-      <h1>Hello React Thingy</h1>
-      
-     
-
+      <div>
+      <PostPage />
+         
       </div>
     );
